06-SDK_RAG/requirements.txt
azure-ai-projects==1.0.0b7
azure-ai-inference[prompts]
azure-identity
azure-search-documents
pandas
python-dotenv
opentelemetry-api
dotenv
azure-monitor-opentelemetry
azure-ai-evaluation
